+++ src/main/java/com/yooksi/betterarchery/common/BetterArchery.java
@@ -15,7 +15,7 @@ public class BetterArchery
 {
 	public static final String MODID = "betterarchery";
 	public static final String NAME = "Better Archery Reborn";
-    public static final String VERSION = "0.1.5";
+    public static final String VERSION = "0.1.6";
 
 	// This is where all our custom items should be listed in-game
 	public static final CreativeTabs tabBetterArchery = new CreativeTabs("BetterArchery")