+++ src/main/java/com/yooksi/betterarchery/network/ClientProxy.java
@@ -1,7 +1,7 @@
 package com.yooksi.betterarchery.network;
 
 import com.yooksi.betterarchery.common.CommonProxy;
-import com.yooksi.betterarchery.common.ResourceLibrary;
+import com.yooksi.betterarchery.init.*;
 
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraftforge.client.model.ModelLoader;
@@ -20,6 +20,6 @@ public class ClientProxy extends CommonProxy
 		// This step is necessary in order to make your block render properly when it is an item (i.e. in the inventory or in your hand or thrown on the ground).
 	    // It must be done on client only, and must be done after the block has been created in Common.preinit().
 		
-		ModelLoader.setCustomModelResourceLocation(com.yooksi.betterarchery.common.ResourceLibrary.REGULAR_BOW, 0, new ModelResourceLocation(ResourceLibrary.REGULAR_BOW.getRegistryName().toString()));
+		ModelLoader.setCustomModelResourceLocation(Items.REGULAR_BOW, 0, new ModelResourceLocation(Items.REGULAR_BOW.getRegistryName().toString()));
 	}
 }